move tests into their correct packages
