+++ packages/protocol/package.json
@@ -15,8 +15,7 @@
   "scripts": {
     "build": "node json-rpc/build-json-rpc-schema.ts && oxfmt --write stagehand.v4.json",
     "test": "vitest run --root ../.. packages/protocol/tests packages/protocol/json-rpc/tests",
-    "test:unit": "vitest run --root ../.. packages/protocol/tests packages/protocol/json-rpc/tests --exclude=packages/protocol/tests/browser-runtime/**",
-    "test:browser": "vitest run --root ../.. packages/protocol/tests/browser-runtime",
+    "test:unit": "vitest run --root ../.. packages/protocol/tests packages/protocol/json-rpc/tests",
     "typecheck": "tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
