+++ packages/sdk-ts/tests/rpcClientTraceContext.test.ts
@@ -1,6 +1,6 @@
 import { ROOT_CONTEXT, TraceFlags, createTraceState, trace } from "@opentelemetry/api";
 import { describe, expect, it } from "vitest";
-import { getTraceContextFields } from "../../../sdk-ts/src/rpcClient.ts";
+import { getTraceContextFields } from "../src/rpcClient.js";
 
 describe("RPC client trace context", () => {
   it("serializes active W3C trace context into JSON-RPC request fields", () => {
